BE - 회원 정보 수정 s3 사진 삭제 조건 수정

+++ src/main/java/com/zerobase/foodlier/global/profile/facade/ProfileFacade.java
@@ -37,7 +37,8 @@ public class ProfileFacade {
                 Objects.isNull(member.getProfileUrl()) ? NO_IMAGE :
                         member.getProfileUrl();
         if (!imageUrl.equals(NO_IMAGE)
-                && Objects.nonNull(member.getProfileUrl())) {
+                && Objects.nonNull(member.getProfileUrl())
+                && !imageUrl.equals(member.getProfileUrl())) {
             s3Service.deleteImage(member.getProfileUrl());
         }
         memberService.updatePrivateProfile(
